+++ sbin/ifconfig/env.h
@@ -7,8 +7,8 @@ const char *getifname(prop_dictionary_t);
 ssize_t getargstr(prop_dictionary_t, const char *, char *, size_t);
 ssize_t getargdata(prop_dictionary_t, const char *, uint8_t *, size_t);
 int getaf(prop_dictionary_t);
-int getifflags(prop_dictionary_t, prop_dictionary_t, unsigned short *);
-const char *getifinfo(prop_dictionary_t, prop_dictionary_t, unsigned short *);
+int getifflags(prop_dictionary_t, prop_dictionary_t, unsigned long int *);
+const char *getifinfo(prop_dictionary_t, prop_dictionary_t, unsigned long int *);
 prop_dictionary_t prop_dictionary_augment(prop_dictionary_t, prop_dictionary_t);
 
 #endif /* _IFCONFIG_ENV_H */
