+++ sys/arch/mips/atheros/dev/aevar.h
@@ -182,7 +182,7 @@ struct ae_softc {
 	struct ae_txsq sc_txfreeq;	/* free Tx descsofts */
 	struct ae_txsq sc_txdirtyq;	/* dirty Tx descsofts */
 
-	short	sc_if_flags;
+	unsigned long int sc_if_flags;
 
 	int	sc_rxptr;		/* next ready RX descriptor/descsoft */
